+++ packages/cliffapp/src/mod.ts
@@ -8,8 +8,7 @@
  * @module
  */
 
-export * from './abstract-cmd.ts';
-export * from './engine.ts';
+export * from './command.ts';
 export * from './logging.ts';
 export * from './run.ts';
 export * from './silent-error.ts';